add Maven, update to JavaFX 16 (from 1.8)
This requires removing the TextFieldCaretControlSkin custom class because the decision to move the caret into javafx.scene.control.skin.TextFieldSkin and keep the caretPath object package private means it is no longer accessible to manual control.
I'll look into ways to style the caret via CSS instead, but not implemented at this point.